refactor: inline ceremonial interfaces
