+++ packages/client/src/signature/MessageSigner.ts
@@ -2,32 +2,36 @@ import { inject, Lifecycle, scoped } from 'tsyringe'
 import { MarkRequired } from 'ts-essentials'
 import { SignatureType, StreamMessage, StreamMessageOptions } from '@streamr/protocol'
 import { Authentication, AuthenticationInjectionToken } from '../Authentication'
-import { LegacySECP256K1Algorithms } from './legacySECP256K1Algorithms'
-import { SECP256K1Algorithms } from './SECP256K1Algorithms'
+import { createSignaturePayload } from './createSignaturePayload'
+import { createLegacySignaturePayload } from './createLegacySignaturePayload'
 
-export interface MessageSignerAlgorithm {
-    sign(opts: MarkRequired<Omit<StreamMessageOptions, 'signature' | 'signatureType'>, 'messageType'>): Promise<Uint8Array>
-}
+type SignerFn = (opts: MarkRequired<Omit<StreamMessageOptions, 'signature' | 'signatureType'>, 'messageType'>) => Promise<Uint8Array>
 
 @scoped(Lifecycle.ContainerScoped)
 export class MessageSigner {
-    private readonly signerAlgorithms = new Map<SignatureType, MessageSignerAlgorithm>()
+    private readonly signers = new Map<SignatureType, SignerFn>()
 
     constructor(@inject(AuthenticationInjectionToken) authentication: Authentication) {
-        this.signerAlgorithms.set(SignatureType.LEGACY_SECP256K1, new LegacySECP256K1Algorithms(authentication)) // TODO: where to register?
-        this.signerAlgorithms.set(SignatureType.SECP256K1, new SECP256K1Algorithms(authentication)) // TODO: where to register?
-        this.signerAlgorithms.set(SignatureType.ERC_1271, new SECP256K1Algorithms(authentication)) // TODO: where to register?
+        this.signers.set(SignatureType.LEGACY_SECP256K1, (opts) => {
+            const payload = createLegacySignaturePayload(opts)
+            return authentication.signWithWallet(payload)
+        }) // TODO: good registration pattern?
+        this.signers.set(SignatureType.SECP256K1, (opts) => {
+            const payload = createSignaturePayload(opts)
+            return authentication.signWithWallet(payload)
+        })
+        this.signers.set(SignatureType.ERC_1271, this.signers.get(SignatureType.SECP256K1)!)
     }
 
     async createSignedMessage(
         opts: MarkRequired<Omit<StreamMessageOptions, 'signature' | 'signatureType'>, 'messageType'>,
         signatureType: SignatureType
     ): Promise<StreamMessage> {
-        const signerAlgorithm = this.signerAlgorithms.get(signatureType)
-        if (signerAlgorithm === undefined) {
-            throw new Error(`Unsupported SignatureType: ${SignatureType}`)
+        const sign = this.signers.get(signatureType)
+        if (sign === undefined) {
+            throw new Error(`Cannot sign message, unsupported signatureType: "${signatureType}"`)
         }
-        const signature = await signerAlgorithm.sign(opts)
+        const signature = await sign(opts)
         return new StreamMessage({
             ...opts,
             signature,
