+++ packages/client/src/signature/SignatureValidator.ts
@@ -1,22 +1,33 @@
 import { Lifecycle, scoped } from 'tsyringe'
 import { SignatureType, StreamMessage, StreamMessageError } from '@streamr/protocol'
 import { ERC1271ContractFacade } from '../contracts/ERC1271ContractFacade'
-import { ERC1271Validator } from './ERC1271Algorithms'
-import { SECP256K1Validator } from './SECP256K1Algorithms'
-import { LegacySECP256K1Validator } from './legacySECP256K1Algorithms'
+import { verifySignature } from '@streamr/utils'
+import { createSignaturePayload } from './createSignaturePayload'
+import { createLegacySignaturePayload } from './createLegacySignaturePayload'
 
-export interface SignatureValidatorAlgorithm {
-    assertSignatureIsValid(streamMessage: StreamMessage): Promise<boolean>
-}
+type ValidatorFn = (streamMessage: StreamMessage) => Promise<boolean>
 
 @scoped(Lifecycle.ContainerScoped)
 export class SignatureValidator {
-    private readonly validatorAlgorithms = new Map<SignatureType, SignatureValidatorAlgorithm>()
+    private readonly validators = new Map<SignatureType, ValidatorFn>()
 
     constructor(erc1271ContractFacade: ERC1271ContractFacade) {
-        this.validatorAlgorithms.set(SignatureType.LEGACY_SECP256K1, new LegacySECP256K1Validator())
-        this.validatorAlgorithms.set(SignatureType.SECP256K1, new SECP256K1Validator())
-        this.validatorAlgorithms.set(SignatureType.ERC_1271, new ERC1271Validator(erc1271ContractFacade))
+        this.validators.set(SignatureType.LEGACY_SECP256K1, async (streamMessage) => {
+            const payload = createLegacySignaturePayload(streamMessage)
+            return verifySignature(streamMessage.getPublisherId(), payload, streamMessage.signature)
+        }) // TODO: good registration pattern?
+        this.validators.set(SignatureType.SECP256K1, async (streamMessage) => {
+            const payload = createSignaturePayload(streamMessage)
+            return verifySignature(streamMessage.getPublisherId(), payload, streamMessage.signature)
+        })
+        this.validators.set(SignatureType.ERC_1271, async (streamMessage) => {
+            const payload = createSignaturePayload(streamMessage)
+            return erc1271ContractFacade.isValidSignature(
+                streamMessage.getPublisherId(),
+                payload,
+                streamMessage.signature
+            )
+        })
     }
 
     /**
@@ -24,13 +35,13 @@ export class SignatureValidator {
      * Resolves if valid, rejects otherwise.
      */
     async assertSignatureIsValid(streamMessage: StreamMessage): Promise<void> {
-        const validatorAlgorithm = this.validatorAlgorithms.get(streamMessage.signatureType)
-        if (validatorAlgorithm === undefined) {
-            throw new Error(`Unsupported SignatureType: ${SignatureType}`)
+        const validate = this.validators.get(streamMessage.signatureType)
+        if (validate === undefined) {
+            throw new Error(`Cannot validate message signature, unsupported signatureType: "${streamMessage.signatureType}"`)
         }
         let success: boolean
         try {
-            success = await validatorAlgorithm.assertSignatureIsValid(streamMessage)
+            success = await validate(streamMessage)
         } catch (err) {
             throw new StreamMessageError(`An error occurred during address recovery from signature: ${err}`, streamMessage)
         }
